+++ keep/api/utils/pagination.py
@@ -2,13 +2,10 @@ from typing import Any
 
 from pydantic import BaseModel
 
-from keep.api.models.alert import (
-    AlertDto,
-    AlertWithIncidentLinkMetadataDto,
-    IncidentDto,
-)
+from keep.api.models.alert import AlertDto, AlertWithIncidentLinkMetadataDto
 from keep.api.models.db.enrichment_event import EnrichmentEvent
 from keep.api.models.db.workflow import *  # pylint: disable=unused-wildcard-importfrom typing import Optional
+from keep.api.models.incident import IncidentDto
 from keep.api.models.workflow import WorkflowDTO, WorkflowExecutionDTO
