+++ tests/test_pagerduty_provider.py
@@ -1,8 +1,8 @@
-import os
-
-import unittest
 import json
-from keep.api.models.alert import IncidentSeverity, IncidentStatus
+import os
+import unittest
+
+from keep.api.models.db.incident import IncidentSeverity, IncidentStatus
 from keep.providers.pagerduty_provider.pagerduty_provider import PagerdutyProvider
 
 
@@ -21,4 +21,3 @@ class TestPagerdutyProvider(unittest.TestCase):
 
 if __name__ == "__main__":
     unittest.main()
-
